Improvements to extension service binding
