+++ src/Gameboard.Api/Common/DontBindForDIAttribute.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace Gameboard.Api.Common;
+
+[AttributeUsage(AttributeTargets.Class | AttributeTargets.Interface)]
+public class DontBindForDIAttribute : Attribute
+{
+}
